chore(style): Replaced explicit `.iter()` calls in for loops with references.
Rust's for loops are actually just syntax sugar over iterators, and as such implicitly call `.into_iter()` on what you provide them. This lets you simply pass the appropriate type to the for loop to get the correct result out, rather than needing to explicitly call the correct iteration function, e.g.

Iterate references, so `letter` is a `&char`:
`for letter in letters_vector.iter()` == `for letter in &letters_vector`

Iterate mutable references, so `letter` is a `&mut char`:
`for letter in letters_vector.iter_mut()` == `for letter in &mut letters_vector`

Iterate owned values and consume the collection, so `letter` is a `char` and `letters_vector` is no longer usable:
`for letter in letters_vector.into_iter()` == `for letter in letters_vector`