perf(letter_display): Allocate memory for string up-front as we know its size and can therefore reduce the number of allocations.
On a string of length 20 this program would probably have to run about a quadrillion times before it saved more time than it took me to write this sentence, but it's good practice nonetheless.

In the general case, there are convenience methods it's good to be aware of like `join` on a slice, or `intersperse` on an Iterator, which do this in its entirety, but interestingly I couldn't get `join` to behave with a `Vec<char>`.